+++ compiler-core/src/analyse/imports.rs
@@ -78,6 +78,9 @@ impl<'a> Importer<'a> {
                 name: import.name.clone(),
                 module_name: module.name.clone(),
                 type_constructors: module.public_type_names(),
+                imported_type_as_value: module
+                    .get_public_value(&import.name)
+                    .map_or_else(|| false, |_| true),
             })?
             .clone()
             .with_location(import.location);
@@ -121,6 +124,9 @@ impl<'a> Importer<'a> {
                     name: import_name.clone(),
                     module_name: module.name.clone(),
                     value_constructors: module.public_value_names(),
+                    imported_value_as_type: module
+                        .get_public_type(import_name)
+                        .map_or_else(|| false, |_| true),
                 });
             }
         };